USB HOST: HID CLASS - получаем координаты от сенсорного экрана

+++ lib/Middlewares/ST/STM32_USB_Host_Library/Class/HID/Inc/usbh_hid_touch.h
@@ -52,8 +52,8 @@ extern "C" {
 
 typedef struct _HID_TOUCH_Info
 {
-  uint8_t              x;
-  uint8_t              y;
+  uint32_t              x;
+  uint32_t              y;
   uint8_t              buttons[3];
 }
 HID_TOUCH_Info_TypeDef;
